Bill S6856 version original - Steve Rhoads
BILL NUMBER: S6856 SPONSOR: RHOADS TITLE OF BILL: An act to amend the penal law, in relation to expanding the offense of unlawfully dealing with a child in the second degree PURPOSE: This legislation increases the age in which a child may enter or remain in certain establishments which serve alcoholic beverages. SUMMARY OF PROVISIONS: Subdivision 1 of section 260.