Bill S6840 version original - Kevin S. Parker
BILL NUMBER: S6840 SPONSOR: PARKER TITLE OF BILL: An act to amend the tax law, in relation to the empire state film production tax credit PURPOSE OR GENERAL IDEA OF BILL: To make documentaries a "Qualified Film" for the purposes of section 24 of the, tax law. SUMMARY OF SPECIFIC PROVISIONS: Paragraph 3 of subdivision (b) of section 24 of the tax law, as added by section 1 of part B of chapter 59 of the laws of 2013, is amended.