Bill S6885A version original - Jack M. Martins
BILL NUMBER: S6885 SPONSOR: MARTINS TITLE OF BILL: An act in relation to authorizing the North Shore Land Alliance, Inc. to file with the county of Nassau assessor an application for a retroac- tive real property tax exemption PURPOSE: To grant the Nassau County Assessor the authority to exempt the North Shore . Land Alliance, Inc. from liability for payment of real property taxes due on the 2023-2024 assessment roll SUMMARY OF PROVISIONS: Section 1.