Bill S6886A version original - Gustavo Rivera
BILL NUMBER: S6886 SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law and the education law, in relation to standing orders for patient care in hospitals PURPOSE OR GENERAL IDEA OF BILL: To enable registered nurses (RNs) to carry out certain procedures with- out a patient-specific order from the attending practitioner. SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds a new § 2803-v-1 to the public health law to permit standing orders for the care of patients in hospitals.