Bill S6892 version original - Iwen Chu
BILL NUMBER: S6892 SPONSOR: CHU TITLE OF BILL: An act to amend the tax law, in relation to enacting the "microbusiness resiliency and growth act" PURPOSE: The purpose of this act is to give microbusinesses in the state a reduced sales tax rate to offset an increase in minimum wage SUMMARY OF PROVISIONS: Section 1 provides a short title for the act. Section 2 amends the tax law, adding a new subdivision (g) by defining the term "microbusiness" as used in the act.