Bill S769B version A - Jeremy Cooney
BILL NUMBER: S769A SPONSOR: COONEY TITLE OF BILL: An act to amend the education law, in relation to creating the profes- sion of nurse anesthetist PURPOSE OF THE BILL: This bill creates a scope of practice for certified registered nurse anesthetists SUMMARY OF SPECIFIC PROVISIONS: Section 1 of this bill creates a new Article 139-A of the education law and defines the scope of practice for certified registered nurse anesth- etist. Section 2 amends the references to who can use the title of nurse.