Bill S6909 version original - James Skoufis
BILL NUMBER: S6909 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the cannabis law, in relation to certification by phar- macists of patients using medical cannabis PURPOSE: To streamline patient certification for medical cannabis and alleviate the administrative burden for medical cannabis providers. SUMMARY OF PROVISIONS: Section 1:Expands the criteria for when a practitioner may certify a patient and reduces the administrative burden for practitioners.