Bill S6910A version original - Gustavo Rivera
BILL NUMBER: S6910 SPONSOR: RIVERA TITLE OF BILL: An act in relation to providing for funding associated with the transi- tion of the pharmacy benefit in the medical assistance program PURPOSE OR GENERAL IDEA OF BILL: To ensure the continuity of funding to support 340B providers and HIV Special Needs Plans (SNPs). SUMMARY OF SPECIFIC PROVISIONS: This bill would require the Governor to include $775 million in funding, adjusted for inflation, in future budgets to support 340B covered enti- ties.