Bill S6928 version original - Joseph A. Griffo
BILL NUMBER: S6928 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the election law, in relation to requiring the county seat in Madison county to have at least one polling place designated for early voting PURPOSE: The purpose of this bill is to ensure that there is an accessible early voting site in Madison County. SUMMARY OF SPECIFIC PROVISIONS: Section 1: amends Election Law to add the county seat of Madison County as a designated early voting polling place.