Bill S3024B version A - Leroy Comrie
BILL NUMBER: S3024A SPONSOR: COMRIE PURPOSE OR GENERAL IDEA OF BILL: This legislation would authorize an innovative approach to land conser- vation that provides environmental, social, and economic benefits to communities in New York. This unique model of land conservation is financially sustainable and contributes to the protection of the envi- ronment while providing an affordable end of life option to communities who need an economic alternative.