Bill S6896 version original - Michelle Hinchey
BILL NUMBER: S6896 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the tax law, in relation to hotel and motel taxes in Ulster county PURPOSE OR GENERAL IDEA OF BILL: This legislation authorizes Ulster County to impose by local law, a tax on hotel and motel occupancy of up to four percent of the charge. SUMMARY OF PROVISIONS: Section 1 amends the Tax Law by amending section 1202-I which authorizes the imposition of a hotel and motel occupancy tax in Ulster County.