Bill S6905 version original - James Skoufis
BILL NUMBER: S6905 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the agriculture and markets law and the environmental conservation law, in relation to prohibiting the use of primates in entertainment acts PURPOSE: This bill enacts the "Primate Protection Act" which prohibits the use of many primates in any type of entertainment act. SUMMARY OF PROVISIONS: Section one: Provides that this act shall be known as the "Primate Protection Act." Section two: Describes legislative findings.