Bill S6915 version original - Timothy M. Kennedy
BILL NUMBER: S6915 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to parking near a stop sign PURPOSE: To allow vehicles to park within 30ft of a stop sign unless posted otherwise. SUMMARY OF PROVISIONS: Section 1. Subparagraph c of paragraph 2 of subdivision (a) of section 1202 of the vehicle and traffic law, as amended by chapter 679 of the laws of 1970, is amended to read as follows: c.