Bill S6933A version original - Lea Webb
BILL NUMBER: S6933 SPONSOR: WEBB TITLE OF BILL: An act to amend the election law, in relation to the terms of office of election commissioners PURPOSE: Relates to the terms of office of commissioners of election. SUMMARY OF PROVISIONS: This act shall amend Subdivision 1 of section 3-202 of the election law. It will extend the term of office of an election commissioner to a four year period statewide beginning January 1st of each odd numbered year.