Bill S2024A version A - Joseph P. Addabbo Jr.
BILL NUMBER: S2024A SPONSOR: ADDABBO TITLE OF BILL: An act to amend the education law and the election law, in relation to school session days PURPOSE: The purpose of this bill is to prohibit schools from being in session on Election Day. Additionally, this bill seeks to provide school districts the ability to require staff attendance or a professional development day on Election Day. Lastly, this bill provides schools with the ability to appeal their designation as a polling place.