Bill S6968 version original - Cordell Cleare
BILL NUMBER: S6968 SPONSOR: CLEARE TITLE OF BILL: An act to direct the department of financial services to study health insurance coverage through the Marketplace PURPOSE: To determine the feasibility of increasing assistance to those who have no health insurance in order to reduce the number of the uninsured. SUMMARY OF PROVISIONS: § 1 Instructs the Department of Financial Services to conduct a study of the healthcare marketplace.