Bill S6995 version original - William Weber
BILL NUMBER: S6995 SPONSOR: WEBER TITLE OF BILL: An act to amend the executive law, in relation to authorizing a county, city, town or village to override a state disaster emergency in all or part of the territorial limits of such local government PURPOSE: Allows a county, city, town or village government to override a state disaster of emergency in all or part of that county, city, town or village's territorial limits. SUMMARY OF PROVISIONS: Section 1. Amends Section 28 of the Executive Law.