Bill S1042A version A - Michelle Hinchey
BILL NUMBER: S1042A SPONSOR: HINCHEY TITLE OF BILL: An act to amend the penal law, in relation to unlawful dissemination or publication of intimate images created by digitization and of sexually explicit depictions of an individual; and to repeal certain provisions of such law relating thereto PURPOSE: This legislation will make it unlawful to disseminate or publicize inti- mate images created by digitization. SUMMARY OF PROVISIONS: Section 1 amends subdivision 1 and 2 of section 245.