Bill S6919A version A - Robert Rolison
BILL NUMBER: S6919A SPONSOR: ROLISON TITLE OF BILL: An act to amend the tax law, in relation to authorizing an occupancy tax in the town of Putnam Valley, in Putnam county; and providing for the repeal of such provisions upon expiration thereof PURPOSE OR GENERAL IDEA OF BILL: Authorizes am occupancy tax in the town of Putnam Valley. SUMMARY OF PROVISIONS: Section 1 - Authorizes the town of Putnam Valley to impose a "hotel" tax and defines the term hotel.