Bill S7003 version original - Brad Hoylman-Sigal
BILL NUMBER: S7003 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the education law, in relation to requiring all school districts to screen all students for dyslexia and dysgraphia PURPOSE: The purpose of this bill is to ensure all students have early screening for dyslexia and dysgraphia. SUMMARY OF SPECIFIC PROVISIONS: Section 1.