Bill S7002B version original - Andrew Gounardes
BILL NUMBER: S7002 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the education law, in relation to limited permits to practice medicine PURPOSE OR GENERAL IDEA OF BILL: To ensure no limitations are placed on the renewal of limited permits to practice medicine. SUMMARY OF PROVISIONS: Section 1 amends the Education Law to outline the duration of limited permits and ensure that no limitations be placed on their biennial renewal. Section 2 states the effective date.