Bill S7022 version original - Timothy M. Kennedy
BILL NUMBER: S7022 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to eligibility for participation in the pre-licensing course internet pilot program PURPOSE: To remove bathers to online instruction for the five-hour pre-licensing course. SUMMARY OF PROVISIONS: Section one of this bill removes a prohibition on drivers holding class DJ and MJ licenses from participating in online pre-licensing courses. Section two provides the effective date.