Bill S6989 version original - James Sanders Jr.
BILL NUMBER: S6989 SPONSOR: SANDERS TITLE OF BILL: An act to amend the lien law, in relation to adding certain notice requirements for enforcing liens on goods in self-storage facilities PURPOSE: This legislation will require greater communication with the occupants of self-storage units whose belongings are at risk of being auctioned. SUMMARY OF SPECIFIC PROVISIONS: Section one amends subparagraph (i) of paragraph (a) and paragraph (c) of 2 subdivision 2 of section 182 of the lien law.