Bill S138A version A - Sean Ryan
BILL NUMBER: S138AREVISED 5/31/23 SPONSOR: RYAN TITLE OF BILL: An act to amend the penal law and the executive law, in relation to certifying instructors in small arms practice PURPOSE: The purpose of this bill is to give the state of New York, and specif- ically the Division of Criminal Justice Services (DCJS), sole authority to develop curricula for certifying instructors in small arms practice. SUMMARY OF PROVISIONS: Section 1 amends subdivision 19 of section 265.