Bill S7013 version original - Neil D. Breslin
BILL NUMBER: S7013 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to establishing a medical loss ratio for dental insurance PURPOSE OR GENERAL IDEA OF BILL: To establish a medical loss ratio for dental insurance and ensure that a minimum amount of the premium is spent on patient care. SUMMARY OF SPECIFIC PROVISIONS: Section 1 establishes a medical loss ratio of 82% for individual and small group dental insurance products.