Bill S5652B version A - Robert Ortt
BILL NUMBER: S5652A SPONSOR: ORTT TITLE OF BILL: An act relating to the dissolution of the Town of Niagara industrial development agency and the disposition of the assets thereof PURPOSE:: To dissolve the Town of Niagara industrial development agency. SUMMARY OF PROVISIONS:: Section 1: The Town of Niagara will now possess the existing records, property, rights, titles, and interests of the Town of Niagara indus- trial development agency.