Bill S7001 version original - George Borrello
BILL NUMBER: S7001 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the uniform city court act, in relation to the selection of certain city court judges in the city of Jamestown PURPOSE OR GENERAL IDEA OF BILL: This bill would provide that both City Court Judges in the city of Jamestown would be elected. SUMMARY OF SPECIFIC PROVISIONS: Section 1 Subparagraph (i) of paragraph 1 of subdivision (e) of section 2104 of the uniform City Court Act, as amended by chapter 455 of the.