Bill S7092 version original - Jessica Ramos
BILL NUMBER: S7092 SPONSOR: RAMOS TITLE OF BILL: An act to amend the tax law, in relation to establishing a tax credit for the payment of union dues PURPOSE: This bill allows taxpayers to claim their union dues as refundable tax credit. SUMMARY OF PROVISIONS: Section 1 amends section 606 of the tax law to allow a tax credit for qualified union dues paid to a labor organization. The definitions of "bargaining representative," "labor organization," and "qualified union dues" are also added.