Bill S6885A version A - Jack M. Martins
BILL NUMBER: S6885AREVISED 5/17/23 SPONSOR: MARTINS TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to the sentence of imprisonment for murder in the first and second degrees PURPOSE: To grant the Nassau County Assessor the authority to exempt the North Shore Land Alliance, Inc. from liability for payment of real property taxes due on the 2022-2023 assessment roll SUMMARY OF PROVISIONS: Section 1.