Bill S6329B version A - Dean Murray
BILL NUMBER: S6329A SPONSOR: MURRAY TITLE OF BILL: An act in relation to authorizing the assessor of the town of Brookha- ven, county of Suffolk to accept an application for a real property tax exemption from De Bethesda Eglise PURPOSE OF BILL: Authorizes the assessor of the town of Brookhaven, county of Suffolk to accept an application for a real property tax exemption from De Bethesda Eglise. SUMMARY OF PROVISIONS: Section 1.