Bill S7040 version original - William Weber
BILL NUMBER: S7040 SPONSOR: WEBER TITLE OF BILL: An act to amend the penal law, in relation to expanding the criteria for a hate crime PURPOSE: Expands the criteria for a hate crime. SUMMARY OF PROVISIONS: Section 1: Subdivision 1 of section 485.05 of the penal law, as amended by chapter 8 or the laws of 2019, is amended. Section 2: Subdivision 2 of section 485.1 of the penal law, as added by chapter 107 of the laws of 2000, is amended.