Bill S1901A version A - Toby Ann Stavisky
BILL NUMBER: S1901A SPONSOR: STAVISKY TITLE OF BILL: An act to amend the criminal procedure law, in relation to victim state- ments at the sentencing of a defendant for a misdemeanor PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to allow victims of certain misdemeanor crimes to request to make a statement at the sentencing of the defendant and to direct the court to accept such request. SUMMARY OF PROVISIONS: Section one provides the short title as "Emma's Law".