Bill S7102 version original - Cordell Cleare
BILL NUMBER: S7102 SPONSOR: CLEARE TITLE OF BILL: An act to repeal paragraph (e) of subdivision 3 of section 2853 of the education law PURPOSE OR GENERAL IDEA OF BILL: To repeal hastily arranged changes in co-locations of Charter Schools that were jammed into the FY 2014 NYS Budget at the last minute. SUMMARY OF PROVISIONS: To repeal paragraph (e) of subdivision 3 of section 2853 of the educa- tion law.