Bill S7146 version original - Kevin S. Parker
BILL NUMBER: S7146 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to radon testing in schools PURPOSE OR GENERAL IDEA OF BILL: This bill would require all school buildings to be tested at least once for radon. Further testing, if any, would be on a schedule to be deter- mined by the Commissioner of Education.