Bill S7117 version original - Timothy M. Kennedy
BILL NUMBER: S7117 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the executive law, in relation to amendments to a tribal-state compact with the Seneca Nation of Indians PURPOSE: To authorize the governor to amend and continue a Nation-State gaming compact with the Seneca Nation of Indians. SUMMARY OF PROVISIONS: Section 1 sets forth legislative intent.