Bill S7145A version original - Kevin S. Parker
BILL NUMBER: S7145 SPONSOR: PARKER TITLE OF BILL: An act to amend the New York city charter, in relation to establishing the office of the youth public advocate in the executive office of the mayor PURPOSE OR GENERAL IDEA OF BILL: To establish a youth public advocate in the executive office of the Mayor of New York City. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the New York City charter to add a new section 20-o. The mayor shall establish an office of the youth public advocate.