Bill S7203 version original - Robert Rolison
BILL NUMBER: S7203 SPONSOR: ROLISON TITLE OF BILL: An act to amend the penal law, in relation to enacting "Laree's law" PURPOSE: Establishes the crime of homicide by sale of an opiate controlled substance. SUMMARY OF PROVISIONS: Section 1 of the bill provides that the short title of this act shall be 1.411t" Section 2 adds a new subdivision 21 to Section 220.00 of the penal law provides the definition of an opiate controlled substance, Section 3 adds a new section 125.