Bill S7205 version original - Leroy Comrie
BILL NUMBER: S7205 SPONSOR: COMRIE TITLE OF BILL: An act to amend the penal law, in relation to the misappropriation of payroll funds PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to establish the crime of misappropriation of payroll funds in New York. SUMMARY OF PROVISIONS: Section one establishes the crime of Misappropriation of Payroll Funds. States that a delay in the transfer of funds when related to an investi- gation cannot be considered a misappropriation.