Bill S7096 version original - Jessica Ramos
BILL NUMBER: S7096 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to awards to employees who prevail in actions on kick-back claims PURPOSE: This bill codifies that if an employee wins a kick-back claim, the employee will be entitled to the wages, supplements and other things of value received. SUMMARY OF SPECIFIC PROVISIONS: Section 1.