Bill S7236A version original - Samra Brouk
BILL NUMBER: S7236 SPONSOR: BROUK TITLE OF BILL: An act to amend the education law, in relation to improving student mental health and mental health literacy in institutions of higher education in the state, and requiring the adoption of related protocols, policies and programs PURPOSE OR GENERAL IDEA OF BILL: To promote policies to increase mental health literacy on campuses and encourage whole. health parity.