Bill S7185 version original - Kevin S. Parker
BILL NUMBER: S7185 SPONSOR: PARKER TITLE OF BILL: An act to require the office for the prevention of domestic violence to conduct a study on domestic violence in the transgender community PURPOSE OF BILL: To establish a study by the office for the prevention of domestic violence. SUMMARY OF PROVISIONS: The office for the prevention of domestic violence shall conduct a study on domestic violence in the transgender community with a focus on domes- tic violence against Black transgender women.