Bill S7192 version original - Kevin S. Parker
BILL NUMBER: S7192 SPONSOR: PARKER TITLE OF BILL: An act to amend the labor law, in relation to including public employ- ment in wage payment provisions PURPOSE: The bill provides a remedy for employees who are not paid by their public employer. SUMMARY OF PROVISIONS: Section 1: Subdivision 3 of section 190 of the labor law, as amended by chapter 281 of the laws of 2002, is amended Section 2: This act shall take effect immediately.