Bill S7206A version original - Andrew Gounardes
BILL NUMBER: S7206 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the retirement and social security law, in relation to a twenty-five year retirement plan for New York city probation officers PURPOSE OR GENERAL IDEA OF BILL: To allow New York City probation officers to participate and become subject to the provisions of a twenty-five year retirement program. SUMMARY OF PROVISIONS: Section one of the bill amends the Retirement and Social Security Law by adding a new section 604-j.