Bill S7188 version original - Kevin S. Parker
BILL NUMBER: S7188 SPONSOR: PARKER TITLE OF BILL: An act to amend the tax law, in relation to establishing the New York state foreign dependent children tax inclusion act PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to allow a resident taxpayer to receive a state income tax standard deduction and tax exemption for the financial support provided to a dependent child or children residing outside of the United States. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Short title.