Bill S7154 version original - Kevin S. Parker
BILL NUMBER: S7154 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to establishing a state youth commission to advise the governor on youth issues in the state of New York PURPOSE OR GENERAL IDEA OF BILL: The purpose of the legislation is to establish a statewide youth commis- sion to advise the governor and the legislature on youth-related issues in New York. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the executive law to establish a state youth commis- sion.