Bill S7134 version original - Jessica Ramos
BILL NUMBER: S7134 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to the classification of certain worksite employees as employees of a professional employer PURPOSE: This bill classifies worksite employees of professional employers as employees of the professional employer for certain insurance purposes. SUMMARY OF SPECIFIC PROVISIONS: Section 1.