Bill S7229A version original - Shelley Mayer
BILL NUMBER: S7229 SPONSOR: MAYER TITLE OF BILL: An act to amend the education law, in relation to absentee ballots for school district elections PURPOSE OR GENERAL IDEA OF BILL: To update requirements for absentee ballot applications for school elections. SUMMARY OF PROVISIONS: Section 1 amends paragraphs a and c section 2018-a of the education law, to clarify valid reasons to obtain an absentee ballot for elections in school districts which do not utilize poll registration.