Bill S2832A version A - Neil D. Breslin
BILL NUMBER: S2832A SPONSOR: BRESLIN TITLE OF BILL: An act to amend the penal law, in relation to wage theft PURPOSE GENERAL IDEA OF BILL: The purpose of this bill is to allow prosecutors to seek stronger penal- ties against employers who steal wages from workers. SUMMARY OF SPECIFIC PROVISIONS: Amend the Penal Law to include wages in the definition of "property." Subsection one of section 155.00 of the Penal Law is amended to read as follows: Penal Law § 155.00 Larceny; definitions of terms.