Bill S7161 version original - John Mannion
BILL NUMBER: S7161 SPONSOR: MANNION TITLE OF BILL: An act to amend the agriculture and markets law, in relation to expand- ing the definition of crops, livestock and livestock products PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to SUMMARY OF PROVISIONS: Section 1 of this bill would establish a definition for purpose bred animal. Section 2 of this bill provides that the effective date shall take effect immediately.