Bill S7155 version original - Kevin S. Parker
BILL NUMBER: S7155 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to establishing a transgender and non-binary peer support grant program PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to establish a grant program for entities that provide transgender and non-binary counseling. SUMMARY OF SPECIFIC PROVISIONS: Section 1: The mental hygiene law is amended by adding a new section 7.