Bill S7184 version original - Kevin S. Parker
BILL NUMBER: S7184 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to designating March seventh as "COVID-19 Remembrance Day" PURPOSE: The purposed bill seeks to recognize and honor those who lost their lives during the COVID-19 pandemic. SUMMARY OF PROVISIONS: This bill would amend section 168-a (3) of the Executive Law to add a day of commemoration, March Seventh of each year to be known as "COVID- 19 Remembrance Day.