Bill S7193 version original - Kevin S. Parker
BILL NUMBER: S7193 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to victims of domestic violence PURPOSE: Authorizes a victim of domestic violence to apply to remove the violent felony offender from deed of co-owned real property. SUMMARY OF PROVISIONS: Section 1. Amends the executive law to add a new section 577.