Bill S7168 version original - Monica Martinez
BILL NUMBER: S7168 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the public health law, in relation to a standardized form for medical exemption from required immunizations PURPOSE: To create a standardized form to be used in schools in order to receive a medical exemption from required vaccinations. SUMMARY OF PROVISIONS: Section 1.