Bill S7219 version original - Michelle Hinchey
BILL NUMBER: S7219 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the not-for-profit corporation law, in relation to exempting the Leeds Hose Company No. 1, Inc. from the requirement that the percentage of non-resident fire department members not exceed forty-five percent of the membership PURPOSE: This bill will allow the Leeds Hose Company No 1, Inc. to expand its membership to include more non-resident volunteer firefighters.