Bill S6130A version A - Kevin S. Parker
BILL NUMBER: S6130A SPONSOR: PARKER TITLE OF BILL: An act to amend the election law, in relation to watchers SUMMARY OF PROVISIONS: Section 1: Subdivision 5 of section 8-500 of the election law, as amended by chapter 418 of the laws of 2019, is amended. Section 2: This act shall take effect immediately. JUSTIFICATION: Attorneys licensed to practice law in New York are not restricted in their practices to any locality or court.