Bill S7177B version original - Nathalia Fernandez
BILL NUMBER: S7177 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the public health law, in relation to increasing the supply of controlled substances a practitioner in a hospital without a full-time pharmacy may dispense to a patient in a hospital emergency room for use off the premises SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends subdivision 2 of section 3342 of the public health law, as 2 amended by chapter 692 of the laws of 1976. Section 2 sets forth the effective date.