Bill S7182 version original - Monica Martinez
BILL NUMBER: S7182 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the penal law, in relation to increasing the penalties for sex trafficking PURPOSE: The purpose of this bill is to recognize the severity of sex trafficking as a crime by increasing its felony classification from class B to class All. SUMMARY OF PROVISIONS: Section 1: The closing paragraph of section 230.