Bill S7304 version original - Jessica Scarcella-Spanton
BILL NUMBER: S7304 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the insurance law, in relation to coverage for electro- lysis procedures to stop the growth of hair, when deemed medically necessary PURPOSE: The purpose of this bill is provide insurance coverage for electrolysis procedures when deemed medically necessary. SUMMARY OF PROVISIONS: Section 1. Subsection (i) of section 3216 of the insurance law is amended.