Bill S7267 version original - Kevin S. Parker
BILL NUMBER: S7267 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to establishing emergency mental health units PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to establish a Fourth emergency response unit for state of emergency mental health calls to 911 SUMMARY OF SPECIFIC PROVISIONS: Section 1 creates a new article 11 under title B of the mental hygiene law. 11.01 provides Declaration of policy and statement of purpose. 11.