Bill S7351 version original - Kevin S. Parker
BILL NUMBER: S7351 SPONSOR: PARKER TITLE OF BILL: An act to amend the criminal procedure law, the family court act and the penal law, in relation to including rape in the third degree, rape in the second degree and rape in the first degree as family offenses PURPOSE OR GENERAL IDEA OF BILL: Relates to including rape in the third degree, rape in the second degree and rape in the first degree as family defense SUMMARY OF SPECIFIC PROVISIONS: Section 1. Subdivision 1 of section 530.