Bill S7273 version original - Kevin S. Parker
BILL NUMBER: S7273 SPONSOR: PARKER TITLE OF BILL: An act to amend the civil service law, in relation to the creation of an annual report on the race and ethnic data of individuals who have taken a civil service examination PURPOSE OR GENERAL IDEA OF THE BILL: This legislation requires the creation of an annual report on the race and ethnic data of individuals who have taken civil service exams. Department of Civil Service has failed to ensure diversity in state government workforce.