Bill S7356 version original - Toby Ann Stavisky
BILL NUMBER: S7356 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to permitting dentists to administer influenza and COVID vaccinations and vaccinations related to a public health emergency SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends section 6601 of the education law to authorize dentists to administer COVID-19, influenza vaccinations, and other vaccinations related to a public health emergency. Section 2 provides for an immediate effective date.