Bill S7290 version original - Joseph A. Griffo
BILL NUMBER: S7290 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the state finance law and the legislative law, in relation to withholding the salary of the governor and members of the legislature if the budget is late until final ratification of the budget PURPOSE: The purpose of this bill is to amend the finance law in relation to salary forfeiture and the passage of a late budget. SUMMARY OF PROVISIONS: Section 1 is the short title.