Bill S7247 version original - John Mannion
BILL NUMBER: S7247 SPONSOR: MANNION TITLE OF BILL: An act to amend the mental hygiene law, in relation to establishing the eating disorders in children task force PURPOSE: The purpose of this bill is to establish an eating disorders in children task force. SUMMARY OF PROVISIONS: Section 1 of this bill would require the commissioner to convene a task force to study and create guidelines, best practices, and a standardized approach for the early screening of eating disorders in children.