Bill S7328 version original - Kevin S. Parker
BILL NUMBER: S7328 SPONSOR: PARKER TITLE OF BILL: An act to amend the retirement and social security law, in relation to prohibiting the investment of certain public funds with the Trump organ- ization PURPOSE OR GENERAL IDEA OF BILL: To prohibit New York State from investing with the Trump organization and affiliated entities. SUMMARY OF SPECIFIC PROVISIONS: Section 1.