Bill S7278 version original - Kevin S. Parker
BILL NUMBER: S7278 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to rental discrimination based upon consumer credit history PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to prohibit the use of consumer credit histo- ry for housing accommodations. SUMMARY OF SPECIFIC PROVISIONS: The bill would make it unlawful to "refuse to rent or lease" housing accommodation because of consumer credit history.