Bill S7258 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S7258 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to authorizing the spouses of certain volunteer firefighters to be issued a distinctive license plate for volunteer firefighters PURPOSE: An act to amend the vehicle and traffic law, in relation to authorizing the spouses of certain volunteer firefighters to be issued a distinctive license plate for volunteer firefighters. SUMMARY OF PROVISIONS: Section 1.